Refactor to move stopdetail data models to :core:model
