+++ core/model/src/main/java/com/chaeny/busoda/model/BusStopDetail.kt
@@ -1,0 +1,6 @@
+package com.chaeny.busoda.model
+
+data class BusStopDetail(
+    val stopName: String,
+    val busInfos: List<BusInfo>
+)
