+++ feature/stopdetail/build.gradle
@@ -52,4 +52,5 @@ dependencies {
     androidTestImplementation libs.androidx.espresso.core
 
     implementation project(":core:data")
+    implementation project(":core:model")
 }
