--- feature/stopdetail/src/main/java/com/chaeny/busoda/stopdetail/BusStopDetail.kt
@@ -1,6 +1,0 @@
-package com.chaeny.busoda.stopdetail
-
-internal data class BusStopDetail(
-    val stopName: String,
-    val busInfos: List<BusInfo>
-)
