+++ feature/stopdetail/src/main/java/com/chaeny/busoda/stopdetail/StopDetailAdapter.kt
@@ -5,6 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.chaeny.busoda.model.BusInfo
 import com.chaeny.busoda.stopdetail.databinding.ListItemBusBinding
 
 internal class StopDetailAdapter : ListAdapter<BusInfo, StopDetailAdapter.BusDetailViewHolder>(BusDetailDiffCallback()) {
